Created two classes for git and ssh commands + Switched Session for Cache
